requirements.txt
djangorestframework==3.2.5
dj-database-url==0.4.2
Django==1.10.6
django-configurations==2.0
gunicorn==19.7.1
psycopg2==2.7.1
whitenoise==3.3.0
django-cors-headers==2.0.2